requirements.txt
# ============================
#  J.A.R.V.I.S - Requirements
#  Author: Moin Raju Chavhan
#  Version: 3.0 (2025)
# ============================

# 🎤 Voice & Speech
pyttsx3==2.91
SpeechRecognition==3.10.4
pyaudio==0.2.14

# 🌐 Web & APIs
requests==2.32.3
geocoder==1.38.1
beautifulsoup4==4.12.3
wikipedia==1.4.0
python-dotenv==1.0.1

# 🤖 AI & NLP
openai==1.13.3

# 💻 Automation & System
pyautogui==0.9.54
psutil==6.0.0
pyjokes==0.6.0

# 🧩 Web Automation (Selenium)
selenium==4.25.0
webdriver-manager==4.0.2

# 📧 Gmail API Integration
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.143.0

# 📂 File Management & Utilities
shutilwhich==1.1.0
pathlib==1.0.1

# 📰 OCR & Computer Vision
opencv-python==4.10.0.84
pytesseract==0.3.13

# 📹 YouTube Automation & Download
pytube==15.0.0
tkintertable==1.3.3

# 🧠 Testing & Debug
pytest==8.3.3
colorama==0.4.6



# 🧰 Optional (Recommended)
playsound==1.3.0
pyperclip==1.9.0
